ghetto bootstrapping via rake, more xen setup

+++ Rakefile
@@ -13,19 +13,27 @@ end
 
 desc 'Ghetto chef bootstrapping'
 task :pre do
-  sh 'sudo apt-get install -y -q rubygems ruby ruby-dev libopenssl-ruby1.8 build-essential vim tree htop git-core'
-  cmd = <<-EoC
-  cd /tmp
-  wget http://production.cf.rubygems.org/rubygems/rubygems-1.3.7.tgz
-  tar zxf rubygems-1.3.6.tgz
-  cd rubygems-1.3.6
-  sudo ruby setup.rb
-  sudo gem update --system
-EoC
-  sh cmd.split * ';'
-  sh 'sudo gem install chef'
-  sh 'sudo mkdir -p /var/chef/cookbooks'
+  unless File.exists?('/var/chef/cookbooks')
+    sh 'sudo apt-get install -y -q rubygems ruby ruby-dev libopenssl-ruby1.8 build-essential vim tree htop git-core'
+    cmd = <<-EoC
+    cd /tmp
+    wget http://production.cf.rubygems.org/rubygems/rubygems-1.3.7.tgz
+    tar zxf rubygems-1.3.6.tgz
+    cd rubygems-1.3.6
+    sudo ruby setup.rb
+    sudo gem update --system
+  EoC
+    sh cmd.split * ';'
+    sh 'sudo gem install chef'
+    sh 'sudo mkdir -p /var/chef'
+    sh 'cd /var/chef; git clone git://github.com/jodell/cookbooks.git'
+  end
 end
+
+task :xen => :pre do
+  sh 'sudo chef-solo -j /var/chef/cookbooks/xen.json'
+end
+
 namespace :cook do
   desc 'default cookbook to install via chef-solo'
   task :default do
